requirements.txt
python-qpid-proton
python-dotenv
